\name{monteCarloProportions-package}
\alias{monteCarloProportions-package}
\alias{monteCarloProportions}
\docType{package}
\title{\packageTitle{monteCarloProportions}}
\description{\packageDescription{monteCarloProportions}}
\section{Package Content}{\packageIndices{monteCarloProportions}}
\author{\packageAuthor{monteCarloProportions}}
\section{Maintainer}{\packageMaintainer{monteCarloProportions}}
\keyword{package}
